third_party/blink/web_tests/external/wpt/scroll-animations/css/at-scroll-timeline-orientation-expected.txt

This is a testharness.js-based test.
PASS Orientation auto behaves as expected
PASS Orientation vertical behaves as expected
PASS Orientation horizontal behaves as expected
PASS Orientation block behaves as expected in horizontal writing-mode
PASS Orientation inline behaves as expected in horizontal writing-mode
FAIL Orientation block behaves as expected in vertical writing-mode assert_equals: expected "125px" but got "175px"
FAIL Orientation inline behaves as expected in vertical writing-mode assert_equals: expected "175px" but got "125px"
Harness: the test ran to completion.
